[REST API] fix metrics tps type bug
Summary: [REST API] fix metrics tps type bug

Test Plan: no

Reviewers: 晓令

Reviewed By: 晓令

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D786006

+++ flink-runtime/src/main/java/org/apache/flink/runtime/rest/handler/util/MutableIOMetrics.java
@@ -208,7 +208,7 @@ public class MutableIOMetrics extends IOMetrics {
 					}
 
 					if (!findTps && metrics.getMetric(MetricNames.IO_NUM_RECORDS_IN_RATE) != null) {
-						long tps = Long.valueOf(metrics.getMetric(MetricNames.IO_NUM_RECORDS_IN_RATE));
+						double tps = Double.valueOf(metrics.getMetric(MetricNames.IO_NUM_RECORDS_IN_RATE));
 						this.tps = Math.max(tps, this.tps);
 					}
 				}
